Add files via upload
BMI is the acronym for Body Mass Index, a parameter adopted by the World Health Organization to calculate the ideal weight of each person.

The index is calculated as follows: the patient's weight is divided by their height squared. An individual is said to have normal weight when the BMI result is between 18.5 and 24.9.

Want to find out your BMI? Enter your weight and height in the fields below and compare with the indexes in the table.